refactor: split region map render core tile and overlay parts (#3878)
Separate terrain tile painting from resource/improvement overlay rendering
so each region_map render submodule stays under the Phase 3 flame size target
without changing map behavior.

Refs #3878
